requirements.txt
pandas==2.0.3
google-analytics-data==0.17.0
